Registrar contract tweaks and user model
Registration already stores per rGROVEf60cd7aac2d8 username and fullname.

The user can be identified by:
	* an username, which could be used as unix login for example,
	* an optional full name.

We so mark fullname as optional, and rename the name field into username
to avoid confusion between name and fullname.

We also provide a method getName to get the full name, or if not specified,
the username.