New CLI command: db:shell
Summary:
This command runs the command line tool to interact
with the database currently selected in configuration,
with the correct arguments also from the configuration.

For example, if MySQL is selected as engine, `./artisan db:shell`
will open a mysql connected to the right host with proper credentials
and the database selected.

This is fully portable to any Laravel application,
and so a good candidate for a standalone package.

Test Plan:
* './artisan db:shell'
* Unit test for LoginException

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D413