Subresource Integrity
Summary:
SRI is a new specification allowing to associate a hash to a resource file.

This allows to trust external CDNs to serve the exact file we computed the hash.
If not, for example because of a MITM attack, browser will block the resource.

Side edit: ' → " as HTML attributes delimiters.

Test Plan: Visit pages and check each resources are correcty loaded.

Reviewers: dereckson

Tags: #security

Differential Revision: http://devcentral.nasqueron.org/D451