New helper function: authurl
Summary:
16ace564 allowed to customize the /auth route.

We provide an authurl() function, with a similar behavior than url().
This is intended to be used in blade templates: {{ authurl('login') }}

This replaces the @authurl Blade directive offered by 16ace564, as
it's substituted by text when compiled in Blade templates, which isn't
really a desired behavior for a configuration variable.

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D521